requirements.txt
appdirs==1.4.0
django-chartit==0.2.9
django-tables2==1.3.0
packaging==16.8
pyparsing==2.1.10
six==1.10.0
XlsxWriter==0.9.6
